Hw4

+++ service/contactsService.js
@@ -1,23 +1,23 @@
 const Contact = require("./schemas/contact");
 
-const getAll = () => {
-  return Contact.find();
+const getAll = (userId) => {
+  return Contact.find({ owner: userId });
 };
 
-const getById = (id) => {
-  return Contact.findOne({ _id: id });
+const getById = (id, userId) => {
+  return Contact.find({ owner: userId, _id: id });
 };
 
-const add = ({ name, phone, email }) => {
-  return Contact.create({ name, phone, email });
+const add = ({ name, phone, email }, owner) => {
+  return Contact.create({ name, phone, email, owner });
 };
 
 const update = (id, fields) => {
-  return Contact.findByIdAndUpdate({ _id: id }, fields, { new: true });
+  return Contact.findOneAndUpdate({ _id: id }, fields, { new: true });
 };
 
-const remove = (id) => {
-  return Contact.findByIdAndRemove({ _id: id });
+const remove = (id, userId) => {
+  return Contact.findByIdAndRemove({ _id: id, owner: userId });
 };
 
 module.exports = {
